backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Libraries
langgraph==0.2.45
anthropic==0.18.1

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
faker==22.0.0
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
